Add files via upload
The first opening frame of the project is project.java which is the login page for the administrator. After clicking the submit button the user goes to the next frame which is Choice.java which basically asks user to choose between Veg,nonveg and beverage section. After choosing one of the option the corresponding veg.java or NON.java gets opened. Here the user gets to see the menu and can choose their order accordingly. After confirming their order the next page user jumps onto is the bill.java which is enables user to choose the mode of payment. On this frame the user gets to see bill and also has a option to check the guidelines for payment and come back to bill.java on button click. Finally, Once the user goes through the guideline he chooses online or offline mode and accordingly Online.java or Offline.java gets opened  .